move sockets-controls into server test driver
